Add spotless google-java check to build
