+++ clients/java/signalr/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'maven'
+    id "com.diffplug.gradle.spotless" version "3.14.0"
 }
 
 group 'com.microsoft.aspnet'
@@ -18,6 +19,15 @@ dependencies {
     testImplementation group: 'junit', name: 'junit', version: '4.12'
     implementation "org.java-websocket:Java-WebSocket:1.3.8"
     implementation 'com.google.code.gson:gson:2.8.5'
+    compile 'com.google.googlejavaformat:google-java-format:1.6'
+}
+
+spotless {
+    java {
+        licenseHeader '// Copyright (c) .NET Foundation. All rights reserved.\n' +
+                '// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.\n' // License header
+        googleJavaFormat()
+    }
 }
 
 task sourceJar(type: Jar) {
