+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/Action1.java
@@ -1,8 +1,8 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public interface Action1<T1> {
-    void invoke(T1 param1);
+  void invoke(T1 param1);
 }
