+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/CallbackMap.java
@@ -1,6 +1,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 import java.util.ArrayList;
@@ -9,26 +9,27 @@ import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class CallbackMap {
-    private ConcurrentHashMap<String, List<ActionBase>> handlers = new ConcurrentHashMap<>();
+  private ConcurrentHashMap<String, List<ActionBase>> handlers = new ConcurrentHashMap<>();
 
-    public void put(String target, ActionBase action) {
-        handlers.computeIfPresent(target, (methodName, handlerList) -> {
-            handlerList.add(action);
-            return handlerList;
+  public void put(String target, ActionBase action) {
+    handlers.computeIfPresent(
+        target,
+        (methodName, handlerList) -> {
+          handlerList.add(action);
+          return handlerList;
         });
-        handlers.computeIfAbsent(target, (ac) -> new ArrayList<>(Arrays.asList(action)));
-    }
+    handlers.computeIfAbsent(target, (ac) -> new ArrayList<>(Arrays.asList(action)));
+  }
 
-    public Boolean containsKey(String key) {
-        return handlers.containsKey(key);
-    }
+  public Boolean containsKey(String key) {
+    return handlers.containsKey(key);
+  }
 
-    public List<ActionBase> get(String key) {
-        return handlers.get(key);
-    }
+  public List<ActionBase> get(String key) {
+    return handlers.get(key);
+  }
 
-    public void remove(String key) {
-        handlers.remove(key);
-    }
+  public void remove(String key) {
+    handlers.remove(key);
+  }
 }
-
