+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HandshakeProtocol.java
@@ -1,20 +1,20 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 import com.google.gson.Gson;
 
 public class HandshakeProtocol {
-    public static Gson gson = new Gson();
-    private static final String RECORD_SEPARATOR = "\u001e";
+  public static Gson gson = new Gson();
+  private static final String RECORD_SEPARATOR = "\u001e";
 
-    public static String createHandshakeRequestMessage(HandshakeRequestMessage message){
-        // The handshake request is always in the JSON format
-        return gson.toJson(message) + RECORD_SEPARATOR;
-    }
+  public static String createHandshakeRequestMessage(HandshakeRequestMessage message) {
+    // The handshake request is always in the JSON format
+    return gson.toJson(message) + RECORD_SEPARATOR;
+  }
 
-    public static HandshakeResponseMessage parseHandshakeResponse(String message){
-        return gson.fromJson(message, HandshakeResponseMessage.class);
-    }
+  public static HandshakeResponseMessage parseHandshakeResponse(String message) {
+    return gson.fromJson(message, HandshakeResponseMessage.class);
+  }
 }
