+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HandshakeResponseMessage.java
@@ -1,16 +1,16 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public class HandshakeResponseMessage {
-    public String error;
+  public String error;
 
-    public HandshakeResponseMessage() {
-        this(null);
-    }
+  public HandshakeResponseMessage() {
+    this(null);
+  }
 
-    public HandshakeResponseMessage(String error) {
-        this.error = error;
-    }
+  public HandshakeResponseMessage(String error) {
+    this.error = error;
+  }
 }
