+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HubConnectionState.java
@@ -1,9 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public enum HubConnectionState {
-    CONNECTED,
-    DISCONNECTED,
+  CONNECTED,
+  DISCONNECTED,
 }
