+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HubException.java
@@ -1,17 +1,16 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public class HubException extends Exception {
-    public HubException() {
-    }
+  public HubException() {}
 
-    public HubException(String errorMessage) {
-        super(errorMessage);
-    }
+  public HubException(String errorMessage) {
+    super(errorMessage);
+  }
 
-    public HubException(String errorMessage, Exception innerException) {
-        super(errorMessage, innerException);
-    }
+  public HubException(String errorMessage, Exception innerException) {
+    super(errorMessage, innerException);
+  }
 }
