+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HubMessage.java
@@ -1,11 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
-/**
- * A base class for hub messages.
- */
+/** A base class for hub messages. */
 public abstract class HubMessage {
-    public abstract HubMessageType getMessageType();
+  public abstract HubMessageType getMessageType();
 }
