+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HubMessageType.java
@@ -1,17 +1,20 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public enum HubMessageType {
-    INVOCATION(1),
-    STREAM_ITEM(2),
-    COMPLETION(3),
-    STREAM_INVOCATION(4),
-    CANCEL_INVOCATION(5),
-    PING(6),
-    CLOSE(7);
+  INVOCATION(1),
+  STREAM_ITEM(2),
+  COMPLETION(3),
+  STREAM_INVOCATION(4),
+  CANCEL_INVOCATION(5),
+  PING(6),
+  CLOSE(7);
 
-    public int value;
-    HubMessageType(int id) { this.value = id; }
+  public int value;
+
+  HubMessageType(int id) {
+    this.value = id;
+  }
 }
