+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/InvocationMessage.java
@@ -1,45 +1,45 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public class InvocationMessage extends HubMessage {
-    int type = HubMessageType.INVOCATION.value;
-    String invocationId;
-    String target;
-    Object[] arguments;
+  int type = HubMessageType.INVOCATION.value;
+  String invocationId;
+  String target;
+  Object[] arguments;
 
-    public InvocationMessage(String target, Object[] args) {
-        this.target = target;
-        arguments = args;
-    }
+  public InvocationMessage(String target, Object[] args) {
+    this.target = target;
+    arguments = args;
+  }
 
-    public String getInvocationId() {
-        return invocationId;
-    }
+  public String getInvocationId() {
+    return invocationId;
+  }
 
-    public void setInvocationId(String invocationId) {
-        this.invocationId = invocationId;
-    }
+  public void setInvocationId(String invocationId) {
+    this.invocationId = invocationId;
+  }
 
-    public String getTarget() {
-        return target;
-    }
+  public String getTarget() {
+    return target;
+  }
 
-    public void setTarget(String target) {
-        this.target = target;
-    }
+  public void setTarget(String target) {
+    this.target = target;
+  }
 
-    public Object[] getArguments() {
-        return arguments;
-    }
+  public Object[] getArguments() {
+    return arguments;
+  }
 
-    public void setArguments(Object[] arguments) {
-        this.arguments = arguments;
-    }
+  public void setArguments(Object[] arguments) {
+    this.arguments = arguments;
+  }
 
-    @Override
-    public HubMessageType getMessageType() {
-        return HubMessageType.INVOCATION;
-    }
+  @Override
+  public HubMessageType getMessageType() {
+    return HubMessageType.INVOCATION;
+  }
 }
