+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/Logger.java
@@ -1,9 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public interface Logger {
-    void log(LogLevel logLevel, String message);
-    void log(LogLevel logLevel, String formattedMessage, Object ... args);
+  void log(LogLevel logLevel, String message);
+
+  void log(LogLevel logLevel, String formattedMessage, Object... args);
 }
