+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/NullLogger.java
@@ -1,12 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public class NullLogger implements Logger {
-    @Override
-    public void log(LogLevel logLevel, String message) { }
+  @Override
+  public void log(LogLevel logLevel, String message) {}
 
-    @Override
-    public void log(LogLevel logLevel, String formattedMessage, Object... args) { }
+  @Override
+  public void log(LogLevel logLevel, String formattedMessage, Object... args) {}
 }
