+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/PingMessage.java
@@ -1,14 +1,14 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public class PingMessage extends HubMessage {
 
-    int type = HubMessageType.PING.value;
+  int type = HubMessageType.PING.value;
 
-    @Override
-    public HubMessageType getMessageType() {
-        return HubMessageType.PING;
-    }
+  @Override
+  public HubMessageType getMessageType() {
+    return HubMessageType.PING;
+  }
 }
