+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/Subscription.java
@@ -1,25 +1,25 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 import java.util.List;
 
 public class Subscription {
-    private CallbackMap handlers;
-    private ActionBase action;
-    private String target;
+  private CallbackMap handlers;
+  private ActionBase action;
+  private String target;
 
-    public Subscription(CallbackMap handlers, ActionBase action, String target) {
-        this.handlers = handlers;
-        this.action = action;
-        this.target = target;
-    }
+  public Subscription(CallbackMap handlers, ActionBase action, String target) {
+    this.handlers = handlers;
+    this.action = action;
+    this.target = target;
+  }
 
-    public void unsubscribe() {
-        List<ActionBase> actions = this.handlers.get(target);
-        if (actions != null){
-            actions.remove(action);
-        }
+  public void unsubscribe() {
+    List<ActionBase> actions = this.handlers.get(target);
+    if (actions != null) {
+      actions.remove(action);
     }
+  }
 }
