+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/Transport.java
@@ -1,12 +1,16 @@
 // Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license
+// information.
 package com.microsoft.aspnet.signalr;
 
 public interface Transport {
-    void start() throws Exception;
-    void send(String message) throws Exception;
-    void setOnReceive(OnReceiveCallBack callback);
-    void onReceive(String message) throws Exception;
-    void stop();
+  void start() throws Exception;
+
+  void send(String message) throws Exception;
+
+  void setOnReceive(OnReceiveCallBack callback);
+
+  void onReceive(String message) throws Exception;
+
+  void stop();
 }
