--- clients/java/signalr/src/test/java/WebSocketTransportTest.java
@@ -1,23 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-import com.microsoft.aspnet.signalr.NullLogger;
-import com.microsoft.aspnet.signalr.Transport;
-import com.microsoft.aspnet.signalr.WebSocketTransport;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-
-public class WebSocketTransportTest {
-
-    @Rule
-    public ExpectedException expectedEx = ExpectedException.none();
-
-    @Test
-    public void WebsocketThrowsIfItCantConnect() throws Exception {
-        expectedEx.expect(Exception.class);
-        expectedEx.expectMessage("There was an error starting the Websockets transport");
-        Transport transport = new WebSocketTransport("www.notarealurl12345.fake", new NullLogger());
-        transport.start();
-    }
-}
